f a a f o

+++ merd2.sh
@@ -57,15 +57,6 @@ if [ -d ./${CONF_PT2} ] ; then
 fi
 
 csleep 2
-
-dqb "BFROE tig"
-csleep 2
-${tig} clone ${branch} https://${CONF_BASEURL}/${CONF_PT2}.git
-[ $? -gt 0 ] && exit
-
-dqb "TGI KO"
-csleep 2
-
 #HUOM.konfiguraatio olisi hyvä jättää hukkaamatta, toisaalta wanhat tauhkat hyvä saada pois sotkemasta
 #(conf kanssa olisi hyvä keksiä jotain josqs, jos ei muuten niin conf.example)
 
@@ -82,13 +73,21 @@ fi
 
 ls -las ./*.conf
 csleep 5
+
+dqb "BFROE tig"
+csleep 2
+${tig} clone ${branch} https://${CONF_BASEURL}/${CONF_PT2}.git
+[ $? -gt 0 ] && exit
+
+dqb "TGI KO"
+csleep 2
 #exit 59
 
 if [ -d  ./${CONF_BASE}.OLD ] ; then
 	for f in $(find ./${CONF_BASE}.OLD -type f -not -name conf) ; do
-		dqb "sudo rm ${f}"
+		#dqb "sudo rm ${f}"
 		sudo rm ${f}
-		csleep 1
+		#csleep 1
 	done
 fi
 
@@ -107,17 +106,17 @@ dqb "NEXT:common_lib"
 csleep 1
 
 if [ -x ./${CONF_LIB} ] ; then
+	. ./${CONF_LIB} #VAIH:josko vähitellen pois kommenteista?
+	enforce_access $(whoami) ${d0}/${CONF_BASE}
+
+else
+	echo "SMTHING WR0NG W/ ${CONF_LIB}"
 	#TODO?:/o/b-juttuja oli kanssa? (mv lähinnä)
 	for d in $(find ${d0} -type f -name "*.sh") ; do chmod 0555 ${d} ; done
-
-	#. ./${CONF_LIB} #TODO:josko vähitellen pois kommenteista?
-	#enforce_access $(whoami) ${d0}/${CONF_BASE}
-
+	
 	sudo chmod 0511 /opt/bin/*.bash
 	#josko nyt jo?
 	for d in $(find ${d0} -type f -name "*.sh") ; do chmod 0555 ${d} ; done
-else
-	echo "SMTHING WR0NG W/ ${CONF_LIB}"
 fi
 
 #210426:vissiin onnistui jo vetämään 7thson-oksan, kertaalleen?
